+++ jina/flow/base.py
@@ -1,5 +1,3 @@
-
-
 import argparse
 import base64
 import copy
@@ -72,10 +70,10 @@ class BaseFlow(JAMLCompatible, ExitStack, metaclass=FlowType):
     _cls_client = Client  #: the type of the Client, can be changed to other class
 
     def __init__(
-        self,
-        args: Optional['argparse.Namespace'] = None,
-        env: Optional[Dict] = None,
-        **kwargs,
+            self,
+            args: Optional['argparse.Namespace'] = None,
+            env: Optional[Dict] = None,
+            **kwargs,
     ):
         super().__init__()
         self._version = '1'  #: YAML version number, this will be later overridden if YAML config says the other way
@@ -195,7 +193,7 @@ class BaseFlow(JAMLCompatible, ExitStack, metaclass=FlowType):
         self._pod_nodes[pod_name] = Pod(args, needs)
 
     def needs(
-        self, needs: Union[Tuple[str], List[str]], name: str = 'joiner', *args, **kwargs
+            self, needs: Union[Tuple[str], List[str]], name: str = 'joiner', *args, **kwargs
     ) -> 'BaseFlow':
         """
         Add a blocker to the Flow, wait until all peas defined in **needs** completed.
@@ -232,11 +230,11 @@ class BaseFlow(JAMLCompatible, ExitStack, metaclass=FlowType):
         return self.needs(name=name, needs=needs, *args, **kwargs)
 
     def add(
-        self,
-        needs: Optional[Union[str, Tuple[str], List[str]]] = None,
-        copy_flow: bool = True,
-        pod_role: 'PodRoleType' = PodRoleType.POD,
-        **kwargs,
+            self,
+            needs: Optional[Union[str, Tuple[str], List[str]]] = None,
+            copy_flow: bool = True,
+            pod_role: 'PodRoleType' = PodRoleType.POD,
+            **kwargs,
     ) -> 'BaseFlow':
         """
         Add a Pod to the current Flow object and return the new modified Flow object.
@@ -291,9 +289,9 @@ class BaseFlow(JAMLCompatible, ExitStack, metaclass=FlowType):
         if 'host' in kwargs:
             m = re.match(_regex_port, kwargs['host'])
             if (
-                kwargs.get('host', __default_host__) != __default_host__
-                and m
-                and 'port_expose' not in kwargs
+                    kwargs.get('host', __default_host__) != __default_host__
+                    and m
+                    and 'port_expose' not in kwargs
             ):
                 kwargs['port_expose'] = m.group(2)
                 kwargs['host'] = m.group(1)
@@ -365,12 +363,12 @@ class BaseFlow(JAMLCompatible, ExitStack, metaclass=FlowType):
         return op_flow
 
     def gather_inspect(
-        self,
-        name: str = 'gather_inspect',
-        uses='_merge_eval',
-        include_last_pod: bool = True,
-        *args,
-        **kwargs,
+            self,
+            name: str = 'gather_inspect',
+            uses='_merge_eval',
+            include_last_pod: bool = True,
+            *args,
+            **kwargs,
     ) -> 'BaseFlow':
         """Gather all inspect Pods output into one Pod. When the Flow has no inspect Pod then the Flow itself
         is returned.
@@ -612,8 +610,9 @@ class BaseFlow(JAMLCompatible, ExitStack, metaclass=FlowType):
 
         return a._pod_nodes == b._pod_nodes
 
+    @property
     @build_required(FlowBuildLevel.GRAPH)
-    def _get_client(self, **kwargs) -> 'Client':
+    def client(self, **kwargs) -> 'Client':
         kwargs.update(self._common_kwargs)
         if 'port_expose' not in kwargs:
             kwargs['port_expose'] = self.port_expose
@@ -709,12 +708,12 @@ class BaseFlow(JAMLCompatible, ExitStack, metaclass=FlowType):
         return '\n'.join(mermaid_graph)
 
     def plot(
-        self,
-        output: Optional[str] = None,
-        vertical_layout: bool = False,
-        inline_display: bool = False,
-        build: bool = True,
-        copy_flow: bool = False,
+            self,
+            output: Optional[str] = None,
+            vertical_layout: bool = False,
+            inline_display: bool = False,
+            build: bool = True,
+            copy_flow: bool = False,
     ) -> 'BaseFlow':
         """
         Visualize the Flow up to the current point
