+++ tests/jinahub/hub_mwu/mwu_encoder.py
@@ -1,16 +1,14 @@
 from typing import Any
 
-import numpy as np
-
-from jina.executors.encoders import BaseEncoder
+from jina import Executor, requests
 
 
-class MWUEncoder(BaseEncoder):
+class MWUEncoder(Executor):
     def __init__(self, greetings: str, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._greetings = greetings
-        self.logger.success(f'look at me! {greetings}')
 
-    def encode(self, content: 'np.ndarray', *args, **kwargs) -> Any:
-        self.logger.info(f'{self._greetings} {content}')
-        return np.random.random([content.shape[0], 3])
+    @requests
+    def encode(self, **kwargs) -> Any:
+        pass
+# rc1
